+++ lib/name_drop/configuration.rb
@@ -7,18 +7,4 @@ module NameDrop
       @access_token = 'NotARealAccessToken'
     end
   end
-
-  class << self
-    def configuration
-      @configuration ||= Configuration.new
-    end
-
-    def configuration=(config)
-      @configuration = config
-    end
-
-    def configure
-      yield(configuration)
-    end
-  end
 end
